% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doAuth}
\alias{doAuth}
\title{Invoke the Authentication Process with Google}
\usage{
doAuth()
}
\description{
This function starts the authentication process with Google.
Note that this functions needs user interaction.
}
\seealso{
\code{\link{getAuth}},\code{\link{loadToken}}
}

