% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAuth.R
\name{getAuth}
\alias{getAuth}
\title{Authentication of R app}
\usage{
getAuth(credentials)
}
\arguments{
\item{credentials}{data.frame with the credential information}
}
\value{
credentials data frame with client token from Google authentication server appended,
which is optionally saved as RData file in .google.auth directory under the current working directory.
}
\description{
getAuth authenticates the R app at the Google authentication server using OAUTH2 and receives the client token.
Usually you need not to run getAuth() explicitly since the whole authentication process is managed by \code{\link{doAuth}}.
}

