% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCredentials.R
\name{getCredentials}
\alias{getCredentials}
\title{Credentials of R app}
\usage{
getCredentials()
}
\value{
Dataframe with the credential information which is cached in working space
and optionally saved as RData file in current working directory.
}
\description{
getCredentials asks the user to provide the credentials (Client Id, Client Secret) from the Api project
and the developer token from the adwords MCC.
Usually you need not to run getCredentials() explicitly since the whole authentication process is managed by \code{\link{doAuth}}.
}

