% Generated by roxygen2 (4.0.2): do not edit by hand
\name{refreshToken}
\alias{refreshToken}
\title{Refresh Access Token}
\usage{
refreshToken(access, credlist)
}
\arguments{
\item{access}{list that contains access token information}

\item{credlist}{list of credentials}
}
\value{
New access token with corresponding time stamp.
}
\description{
refreshToken returns a new valid access token. The access token deprecates after one hour and has to updated with the refresh token.
Usually you need not to run refreshToken() explicitly since the whole authentication process is managed by \code{\link{doAuth}}.
}

