% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Show available Adwords Reports}
\usage{
reports(apiVersion = "201603")
}
\arguments{
\item{apiVersion}{Supports 201509, 201601 or 201603. Default is 201603.}
}
\value{
Available report types.
}
\description{
reports provides an overview of all available Adwords report types. The report type is specified in \code{\link{statement}}.
}

