% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformDataDT.R
\name{transformDataDT}
\alias{transformDataDT}
\title{Read data into R data.table}
\usage{
transformDataDT(data, report)
}
\arguments{
\item{data}{Raw csv data from Adwords API.}

\item{report}{filename with report field metadata}
}
\value{
data.table with the Adwords Data.
}
\description{
Reads the csv data file received from the Adwords API into a data.table
 The function is used inside \code{\link{transformData}} when the data.table package is available.
}

