% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformDataNoDT.R
\name{transformDataNoDT}
\alias{transformDataNoDT}
\title{Read data into R dataframe}
\usage{
transformDataNoDT(data, report)
}
\arguments{
\item{data}{Raw csv data from Adwords API.}

\item{report}{filename with report field metadata}
}
\value{
Dataframe with the Adwords Data.
}
\description{
Reads the csv data file received from the Adwords API into a dataframe.
 The function is used inside \code{\link{transformData}} when the data.table package is not available.
}

