% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/changeNames.R
\name{changeNames}
\alias{changeNames}
\title{Change Names of Attributes/variables}
\usage{
changeNames(data)
}
\arguments{
\item{data}{Transformed dataframe}
}
\value{
New column names of dataframe.
}
\description{
Converts the default display names into nicer or more practical names.
}

