% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doAuth.R
\name{doAuth}
\alias{doAuth}
\title{Invoke the Authentication Process with Google}
\usage{
doAuth(save = T)
}
\arguments{
\item{save}{logical denotes whether authentication information should be saved on disk. Defaults to TRUE.}
}
\description{
This function starts the authentication process with Google.
Note that this functions needs user interaction.
}
\seealso{
\code{\link{getAuth}},\code{\link{loadToken}}
}

