% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doAuth.R
\name{doAuth}
\alias{doAuth}
\title{Invoke the Authentication Process with Google}
\usage{
doAuth(save = T, localname = "default")
}
\arguments{
\item{save}{logical denotes whether authentication information should be saved on disk. Defaults to TRUE.}

\item{localname}{alias to save the auth token under. Defaults to "default".}
}
\description{
This function starts the authentication process with Google.
Note that this functions needs user interaction.
}
\seealso{
\code{\link{getCredentials}},\code{\link{getAuth}},\code{\link{loadToken}}
}

