% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Get Metrics/Attributes of specified Report}
\usage{
metrics(report = "ACCOUNT_PERFORMANCE_REPORT", apiVersion = "201509")
}
\arguments{
\item{report}{Report type}

\item{apiVersion}{Supports 201509 or 201506. Default is 201509.}
}
\value{
List of available metrics/attributes.
}
\description{
metrics provides an overview of all available metrics/attributes for a specified report type.
}

