% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Get Metrics/Attributes of specified Report}
\usage{
metrics(report = "ACCOUNT_PERFORMANCE_REPORT", apiVersion = "201506")
}
\arguments{
\item{report}{Report type}

\item{apiVersion}{Supports 201506 or 201502. Default is 201506.}
}
\value{
List of available metrics/attributes.
}
\description{
metrics provides an overview of all available metrics/attributes for a specified report type.
}

