\name{metrics}
\alias{metrics}
\title{Get Metrics/Attributes of specified Report}
\usage{
metrics(report = "ACCOUNT_PERFORMANCE_REPORT")
}
\arguments{
  \item{report}{Report type}
}
\value{
List of available metrics/attributes.
}
\description{
metrics provides an overview of all available
metrics/attributes for a specified report type.
}
\examples{
metrics(report="ACCOUNT_PERFORMANCE_REPORT")
}

