% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metrics}
\alias{metrics}
\title{Get Metrics/Attributes of specified Report}
\usage{
metrics(report = "ACCOUNT_PERFORMANCE_REPORT", apiVersion = "201502")
}
\arguments{
\item{report}{Report type}

\item{apiVersion}{Supports 201502 or 201409. Default is 201502.}
}
\value{
List of available metrics/attributes.
}
\description{
metrics provides an overview of all available metrics/attributes for a specified report type.
}

