% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshToken.R
\name{refreshToken}
\alias{refreshToken}
\title{Refresh Access Token}
\usage{
refreshToken(google_auth)
}
\arguments{
\item{google_auth}{list of credentials and access token}
}
\value{
New access token with corresponding time stamp.
}
\description{
refreshToken returns a new valid access token. The access token deprecates after one hour and has to updated with the refresh token.
Usually you need not to run refreshToken() explicitly since the whole authentication process is managed by \code{\link{doAuth}}.
}
