% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Show available Adwords Reports}
\usage{
reports(apiVersion = "201502")
}
\arguments{
\item{apiVersion}{Supports 201502 or 201409. Default is 201502.}
}
\value{
Available report types.
}
\description{
reports provides an overview of all available Adwords report types. The report type is specified in \code{\link{statement}}.
}

