% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Show available Adwords Reports}
\usage{
reports(apiVersion = "201809")
}
\arguments{
\item{apiVersion}{Supports 201809, 201806 and 201802. Defaults to 201809.}
}
\value{
Available report types.
}
\description{
reports provides an overview of all available Adwords report types. The report type is specified in \code{\link{statement}}.
}
