% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Show available Adwords Reports}
\usage{
reports(apiVersion = "201708")
}
\arguments{
\item{apiVersion}{Supports 201702, 201708. Defaults to 201708.}
}
\value{
Available report types.
}
\description{
reports provides an overview of all available Adwords report types. The report type is specified in \code{\link{statement}}.
}
