% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reports.R
\name{reports}
\alias{reports}
\title{Show available Adwords Reports}
\usage{
reports(apiVersion = "201605")
}
\arguments{
\item{apiVersion}{Supports 201509, 201601, 201603 and 201605. Default is 201605.}
}
\value{
Available report types.
}
\description{
reports provides an overview of all available Adwords report types. The report type is specified in \code{\link{statement}}.
}

